requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-yasg
pillow
